NetAddr v0.0.7 NetAddr

General functions for network address parsing and manipulation, with support for addresses of arbitrary size.

Types

t ::
  NetAddr.Generic.t |
  NetAddr.IPv4.t |
  NetAddr.IPv6.t |
  NetAddr.MAC_48.t

Functions

address(netaddr)

Specs

address(NetAddr.t) :: String.t

Returns a human-readable string for the address part of netaddr.

Examples

iex> NetAddr.address NetAddr.ip("192.0.2.1/24")
"192.0.2.1"

iex> NetAddr.address NetAddr.netaddr(<<1, 2, 3, 4, 5>>)
"0x0102030405"

address_length(netaddr)

Specs

address_length(NetAddr.t) :: non_neg_integer

Return the address length of netaddr.

Examples

iex> NetAddr.address_length NetAddr.ip("192.0.2.1/24")
24

address_length(netaddr, new_length)

Specs

address_length(NetAddr.t, pos_integer) :: NetAddr.t

Returns a new NetAddr.t/0 with the address part of netaddr and the given address length.

Examples

iex> NetAddr.ip("192.0.2.1/24") |> NetAddr.address_length(22)
%NetAddr.IPv4{address: <<192, 0, 2, 1>>, length: 22}

apply_mask(address, mask)

Specs

apply_mask(binary, binary) :: binary

Bitwise ANDs two address binaries, returning the result.

Examples

iex> NetAddr.apply_mask <<192, 0, 2, 1>>, <<255, 255, 255, 0>>
<<192, 0, 2, 0>>

aton(address)

Specs

aton(binary) :: non_neg_integer

Converts address to a decimal.

Examples

iex> NetAddr.aton <<192,0,2,1>>
3221225985

iex> NetAddr.aton <<254, 128, 0, 0, 193, 0, 0, 0, 0, 0, 0, 0, 0, 0, 196, 1>>
338288524986991696549538495105230488577

iex> NetAddr.aton(<<1,2,3,4,5>>)
4328719365

broadcast(ipv4_netaddr)

Specs

broadcast(NetAddr.IPv4.t) :: String.t

Returns a human-readable string for the last address in ipv4_netaddr.

Examples

iex> NetAddr.broadcast NetAddr.ip("192.0.2.1/24")
"192.0.2.255"

contains?(netaddr1, netaddr2)

Specs

contains?(NetAddr.t, NetAddr.t) :: boolean

Returns true if netaddr2 is a subset of netaddr1, up to equality. Otherwise, returns false.

Examples

iex> NetAddr.ip("192.0.2.0/24") |> NetAddr.contains?(NetAddr.ip("192.0.2.0/25"))
true

iex> NetAddr.ip("192.0.2.0/24") |> NetAddr.contains?(NetAddr.ip("192.0.2.0/24"))
true

iex> NetAddr.ip("192.0.2.0/25") |> NetAddr.contains?(NetAddr.ip("192.0.2.0/24"))
false

iex> NetAddr.ip("192.0.2.0/25") |> NetAddr.contains?(NetAddr.ip("192.0.2.128/25"))
false

first_address(netaddr)

Returns a new NetAddr.t/0 with the first address in netaddr.

Examples

iex> NetAddr.first_address NetAddr.ip("192.0.2.1/24")
%NetAddr.IPv4{address: <<192, 0, 2, 0>>, length: 24}

ip(ip_string)

Specs

ip(String.t) :: NetAddr.IPv4.t | NetAddr.IPv6.t

Parses ip_string as an IPv4/IPv6 address or CIDR, returning a NetAddr.IPv4.t/0 or NetAddr.IPv6.t/0 as appropriate.

Examples

iex> NetAddr.ip("192.0.2.1")
%NetAddr.IPv4{address: <<192,0,2,1>>, length: 32}

iex> NetAddr.ip("192.0.2.1/24")
%NetAddr.IPv4{address: <<192,0,2,1>>, length: 24}

iex> NetAddr.ip("fe80::c101")
%NetAddr.IPv6{address: <<0xfe,0x80,0::12*8,0xc1,0x01>>, length: 128}

iex> NetAddr.ip("fe80::c101/64")
%NetAddr.IPv6{address: <<0xfe,0x80,0::12*8,0xc1,0x01>>, length: 64}

ip(ip_address_string, ip_mask_string_or_length)

Specs

ip(String.t, String.t) ::
  NetAddr.IPv4.t |
  NetAddr.IPv6.t

ip(String.t, non_neg_integer) ::
  NetAddr.IPv4.t |
  NetAddr.IPv6.t

Parses ip_address_string with the given address length or ip_mask_string.

Examples

iex> NetAddr.ip "0.0.0.0", 0
%NetAddr.IPv4{address: <<0, 0, 0, 0>>, length: 0}

iex> NetAddr.ip "192.0.2.1", 24
%NetAddr.IPv4{address: <<192, 0, 2, 1>>, length: 24}

iex> NetAddr.ip "192.0.2.1", "255.255.255.0"
%NetAddr.IPv4{address: <<192, 0, 2, 1>>, length: 24}

iex> NetAddr.ip "fe80:0:c100::c401", 64
%NetAddr.IPv6{address: <<254, 128, 0, 0, 193, 0, 0, 0, 0, 0, 0, 0, 0, 0, 196, 1>>, length: 64}

last_address(netaddr)

Returns a new NetAddr.t/0 with the last address in netaddr.

Examples

iex> NetAddr.last_address NetAddr.ip("192.0.2.1/24")
%NetAddr.IPv4{address: <<192, 0, 2, 255>>, length: 24}

length_to_mask(address_length, mask_length_in_bytes)

Specs

length_to_mask(non_neg_integer, pos_integer) :: binary

Converts address_length to an address mask binary.

Examples

iex> NetAddr.length_to_mask(30, 4)
<<255, 255, 255, 252>>

iex> NetAddr.length_to_mask(64, 16)
<<255, 255, 255, 255, 255, 255, 255, 255, 0, 0, 0, 0, 0, 0, 0, 0>>

iex> NetAddr.length_to_mask(37, 6)
<<255, 255, 255, 255, 248, 0>>

mac_48(mac_string)

Specs

mac_48(binary) :: NetAddr.MAC_48.t

Parses mac_string, returning a NetAddr.MAC_48.t/0.

For manifest reasons, the corresponding parser may be robust to the point of returning incorrect results. Caveat emptor.

Examples

iex> NetAddr.mac_48 "01:23:45:67:89:AB"
%NetAddr.MAC_48{address: <<0x01,0x23,0x45,0x67,0x89,0xab>>, length: 48}

iex> NetAddr.mac_48 "01-23-45-67-89-AB"
%NetAddr.MAC_48{address: <<0x01,0x23,0x45,0x67,0x89,0xab>>, length: 48}

iex> NetAddr.mac_48 "0123456789aB"
%NetAddr.MAC_48{address: <<0x01,0x23,0x45,0x67,0x89,0xab>>, length: 48}

iex> NetAddr.mac_48 "01 23 45 67 89 AB"
%NetAddr.MAC_48{address: <<0x01,0x23,0x45,0x67,0x89,0xab>>, length: 48}

iex> NetAddr.mac_48 "\"0fF:33-C0.Ff   33 \""
%NetAddr.MAC_48{address: <<0x0f, 0xf, 0x33, 0xc0, 0xff, 0x33>>, length: 48}

iex> NetAddr.mac_48 "1:2:3:4:5:6"
%NetAddr.MAC_48{address: <<1,2,3,4,5,6>>, length: 48}

iex> NetAddr.mac_48 "01-23-45-67-89-ag"
%NetAddr.MAC_48{address: <<0x01,0x23,0x45,0x67,0x89,0xa>>, length: 48}

iex> NetAddr.mac_48 "123456789aB"
%NetAddr.MAC_48{address: <<0x12,0x34,0x56,0x78,0x9a,0xb>>, length: 48}

mask_to_length(address_mask)

Specs

mask_to_length(binary) :: non_neg_integer

Converts address_mask to an address length.

Examples

iex> NetAddr.mask_to_length(<<255,255,248,0>>)
21

netaddr(address)

Specs

netaddr(binary) :: NetAddr.t

Constructs a NetAddr.t/0 struct given a network address binary.

Examples

iex> NetAddr.netaddr(<<1, 2, 3, 4, 5, 6>>)
%NetAddr.MAC_48{address: <<1, 2, 3, 4, 5, 6>>, length: 48}

iex> NetAddr.netaddr(<<1, 2, 3, 4, 5>>)
%NetAddr.Generic{address: <<1, 2, 3, 4, 5>>, length: 40}

netaddr(address, address_length)

Specs

netaddr(binary, non_neg_integer) :: NetAddr.t

Constructs a NetAddr.t/0 struct given a network address binary and an address length.

netaddr(address, address_length, size_in_bytes)

Specs

netaddr(binary, non_neg_integer, pos_integer) :: NetAddr.Generic.t

Explicitly constructs a NetAddr.Generic.t/0 struct.

Examples

iex> NetAddr.netaddr(<<1, 2, 3, 4, 5, 6>>, 48, 6)
%NetAddr.Generic{address: <<1, 2, 3, 4, 5, 6>>, length: 48}

iex> NetAddr.netaddr(<<1, 2, 3, 4, 5>>, 48, 6)
%NetAddr.Generic{address: <<0, 1, 2, 3, 4, 5>>, length: 48}

netaddr_to_range(netaddr)

Specs

netaddr_to_range(NetAddr.t) :: Range.t

Converts a NetAddr.t/0 to a Range.t/0.

Examples

iex> NetAddr.ip("198.51.100.0/24") |> NetAddr.netaddr_to_range
3325256704..3325256959

netaddr_to_string(netaddr)

Specs

netaddr_to_string(NetAddr.t) :: String.t

Returns a human-readable CIDR or pseudo-CIDR for netaddr.

This is like NetAddr.prefix/1 except host bits are not set to zero. All String.Chars implementations call this function.

Examples

iex> NetAddr.netaddr_to_string %NetAddr.Generic{address: <<1, 2, 3, 4, 5>>, length: 32}
"0x0102030405/32"

network(netaddr)

Specs

network(NetAddr.t) :: String.t

Returns a human-readable string for the first address in netaddr.

Examples

iex> NetAddr.network NetAddr.ip("192.0.2.1/24")
"192.0.2.0"

ntoa(decimal, size_in_bytes)

Specs

ntoa(non_neg_integer, pos_integer) :: binary

Converts decimal to an address.

Examples

iex> NetAddr.ntoa 3221225985, 4
<<192, 0, 2, 1>>

iex> NetAddr.ntoa 338288524986991696549538495105230488577, 16
<<254, 128, 0, 0, 193, 0, 0, 0, 0, 0, 0, 0, 0, 0, 196, 1>>

iex> NetAddr.ntoa 4328719365, 5
<<1, 2, 3, 4, 5>>

prefix(netaddr)

Specs

prefix(NetAddr.t) :: String.t

Returns a human-readable CIDR for the first address in netaddr.

Examples

iex> NetAddr.prefix %NetAddr.IPv4{address: <<192, 0, 2, 1>>, length: 24}
"192.0.2.0/24"

iex> NetAddr.prefix %NetAddr.Generic{address: <<1, 2, 3, 4, 5>>, length: 32}
"0x0102030400/32"

range_to_netaddr(range, size_in_bytes)

Specs

range_to_netaddr(Range.t, pos_integer) :: NetAddr.t

Converts range to a NetAddr.t/0 given an address size hint.

Examples

iex> NetAddr.range_to_netaddr 3325256704..3325256959, 4
%NetAddr.IPv4{address: <<198, 51, 100, 0>>, length: 24}

subnet_mask(ipv4_netaddr)

Specs

subnet_mask(NetAddr.IPv4.t) :: String.t

Returns a human-readable address mask for ipv4_netaddr.

Examples

iex> NetAddr.subnet_mask %NetAddr.IPv4{address: <<192, 0, 2, 1>>, length: 24}
"255.255.255.0"